+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix: fast HttpClient timeout fixed
 
 **2.0.54.0 (2022-04-26)**
+* ***Contains critical bug and should not be used. Use version 2.0.54.1. ***
 * Separate HttpClient for critical and other requests
 * Added configuration option for fast HttpClient in OperationManager (default timeout 5s)
 * Extended SportsInfoManager with getTimelineEvents